# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Surveys:
    properties:
      created_at:
        type: string
      id:
        type: integer
      questions:
        type: string
      survey_description:
        type: string
      survey_name:
        type: string
      survey_slug:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UserAnswers:
    properties:
      answers:
        additionalProperties: true
        type: object
      created_at:
        type: string
      id:
        type: integer
      survey_id:
        type: integer
      survey_slug:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:5050
info:
  contact:
    email: support@swagger.io
    name: Arly API Support
    url: http://www.swagger.io/support
  description: This is the Arly API server documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Arly API
  version: "1.0"
paths:
...
  /api/surveys:
    get:
      consumes:
      - application/json
      description: Get a list of all surveys.
      produces:
      - application/json
      responses:
        "200":
          description: List of surveys
          schema:
            items:
              $ref: '#/definitions/models.Surveys'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Fetch all surveys
      tags:
      - Surveys
    post:
      consumes:
      - application/json
      description: Admins can create a new survey with embedded questions.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Survey data
        in: body
        name: survey
        required: true
        schema:
          $ref: '#/definitions/models.Surveys'
      produces:
      - application/json
      responses:
        "201":
          description: Survey created successfully
          schema:
            $ref: '#/definitions/models.Surveys'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new survey
      tags:
      - Surveys
  /api/surveys/{id}:
    delete:
      consumes:
      - application/json
      description: Admins can delete a survey by ID.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Survey ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Survey deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Survey not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a survey
      tags:
      - Surveys
    get:
      consumes:
      - application/json
      description: Get a specific survey along with its embedded questions.
      parameters:
      - description: Survey ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Survey details with questions
          schema:
            $ref: '#/definitions/models.Surveys'
        "404":
          description: Survey not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Fetch a survey with its embedded questions
      tags:
      - Surveys
    put:
      consumes:
      - application/json
      description: Admins can update a survey's details including questions.
      parameters:
      - description: Survey ID
        in: path
        name: id
        required: true
        type: integer
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated survey data
        in: body
        name: survey
        required: true
        schema:
          $ref: '#/definitions/models.Surveys'
      produces:
      - application/json
      responses:
        "200":
          description: Survey updated successfully
          schema:
            $ref: '#/definitions/models.Surveys'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Survey not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing survey
      tags:
      - Surveys
  /api/surveys/{id}/responses:
    get:
      consumes:
      - application/json
      description: Admins can fetch all responses for a specific survey.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Survey ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of survey responses
          schema:
            items:
              $ref: '#/definitions/models.UserAnswers'
            type: array
        "404":
          description: Survey not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fetch all responses for a survey
      tags:
      - Surveys
    post:
      consumes:
      - application/json
      description: Users can submit their responses to a survey.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Survey ID
        in: path
        name: id
        required: true
        type: integer
      - description: User's survey response
        in: body
        name: response
        required: true
        schema:
          $ref: '#/definitions/models.UserAnswers'
      produces:
      - application/json
      responses:
        "201":
          description: Response submitted successfully
          schema:
            $ref: '#/definitions/models.UserAnswers'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Survey not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit a user's survey response
      tags:
      - Surveys
  /api/surveys/{id}/responses/{user_id}:
    get:
      consumes:
      - application/json
      description: Get a specific user's response to a survey.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Survey ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User's survey response
          schema:
            $ref: '#/definitions/models.UserAnswers'
        "404":
          description: Response not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fetch a user's response for a survey
      tags:
      - Surveys
  /api/surveys/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get a specific survey by its slug.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Survey slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Survey details with questions
          schema:
            $ref: '#/definitions/models.Surveys'
        "404":
          description: Survey not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fetch a survey by slug
      tags:
      - Surveys
...
schemes:
- http
swagger: "2.0"